+++ database/migrations/2022_06_17_125807_create_criteria_table.php
@@ -18,7 +18,7 @@ class CreateCriteriaTable extends Migration
             $table->timestamps();
             $table->string("criteria_name");
             $table->string("attribute");
-            $table->float("weight");
+            $table->double("weight");
         });
     }
